+++ src/context/colorScheme.tsx
@@ -1,3 +1,8 @@
+'use client'
+
+import { getCookie, setCookie } from "@/lib/cookies"
+import { createContext, Dispatch, ReactNode, SetStateAction, useContext, useEffect, useState } from "react"
+
 export const colorScheme = {
     navbar: {
         red:{
@@ -127,4 +132,100 @@ export const colorScheme = {
             }
         }
     }
+}
+
+export type AccentColors = 'red'|'green'|'blue'
+
+type colorContextType = {
+    effectiveTheme: 'light'|'dark';
+    theme: 'light'|'default'|'dark';
+    accentColor: 'red'|'green'|'blue';
+    setEffectiveTheme: Dispatch<SetStateAction<"light" | "dark">>;
+    setTheme: Dispatch<SetStateAction<"light" | "dark" | "default">>;
+    setAccentColor: Dispatch<SetStateAction<AccentColors>>;
+}
+
+const colorContext = createContext<colorContextType>({
+    effectiveTheme: 'dark',
+    theme: 'dark',
+    accentColor: 'blue',
+    setEffectiveTheme: function (value: SetStateAction<"light" | "dark">): void {},
+    setTheme: function (value: SetStateAction<"light" | "dark" | "default">): void {},
+    setAccentColor: function (value: SetStateAction<AccentColors>): void {}
+});
+
+export const ColorContextProvider= ({children}:{children:ReactNode})=>{
+
+    const [effectiveTheme, setEffectiveTheme] = useState<'light'|'dark'>('light');
+    const [theme, setTheme] = useState<'dark'|'light'|'default'>('light');
+    const [accentColor, setAccentColor] = useState<AccentColors>('blue');
+
+    //Checks Theme cookie on 1st render
+    useEffect(()=>{
+        let cookie = getCookie('ColorScheme').cookie
+        let cookie_theme, cookie_accentColor
+
+        if(cookie){
+            cookie_theme = cookie.split('; ')[0].split(': ')[1] as 'light'|'dark'|'default'
+            cookie_accentColor = cookie.split('; ')[1].split(': ')[1] as AccentColors
+        }
+        if(cookie_theme && cookie_accentColor){
+            setTheme(cookie_theme)
+            setAccentColor(cookie_accentColor)
+        }else setCookie('ColorScheme',`Theme: ${theme}; AccentColor: ${accentColor}`)
+
+    },[])
+    //Checks for Theme changes
+    useEffect(()=>{
+        let cookie = getCookie('ColorScheme').cookie
+        let cookie_theme, cookie_accentColor
+        if(cookie){
+            cookie_theme = cookie.split('; ')[0].split(': ')[1] as 'light'|'dark'|'default'
+            cookie_accentColor = cookie.split('; ')[1].split(': ')[1] as AccentColors
+        }
+        if (cookie_theme!== theme || cookie_accentColor!== accentColor)
+            setCookie('ColorScheme', `Theme: ${theme}; AccentColor: ${accentColor}`)
+    },[theme,accentColor])
+
+    useEffect(()=>{
+        const browserThemeChecker = (event:any)=>{
+            if (theme === 'default') {
+                if (event.matches){
+                    setEffectiveTheme('dark')
+                }
+                else{
+                    setEffectiveTheme('light')
+                }
+            }
+            else {
+                setEffectiveTheme(theme=== 'light'? 'light':'dark')
+            }
+        }
+        const mediaQuery = window.matchMedia('(prefers-color-scheme: dark)')
+
+        //initial Run
+        browserThemeChecker(window.matchMedia('(prefers-color-scheme: dark)'))
+        
+        mediaQuery.addEventListener('change',browserThemeChecker)
+        return ()=> {
+            mediaQuery.removeEventListener('change',browserThemeChecker)
+        }
+
+    },[theme])
+
+    
+    const value:colorContextType = {
+        effectiveTheme: effectiveTheme,
+        theme: theme,
+        accentColor: accentColor,
+        setEffectiveTheme: setEffectiveTheme,
+        setTheme: setTheme,
+        setAccentColor: setAccentColor
+    }
+    return <colorContext.Provider value={value}>{children}</colorContext.Provider>
+}
+
+export const useColorContext = ()=>{
+    const {effectiveTheme, setEffectiveTheme, theme, setTheme, accentColor, setAccentColor} = useContext(colorContext);
+    return { effectiveTheme, theme, setTheme, accentColor, setAccentColor }
 }